% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullclinesPhaseArrows.R
\name{nullclines_Math1808}
\alias{nullclines_Math1808}
\title{Create nullclines for the phase plan}
\usage{
nullclines_Math1808(fun, xlims, ylims, resol = 250, add = F, parms = NULL)
}
\description{
This function creates nullclines within a phase plan.
It assumes that you have a differential equation or system of DEs
of the following form:
morrislecar=function(V,W,parms){
g_L=2; g_Ca=4; g_K=8; phi=0.04;
cap=20; V_1=-1.2; V_2=18; V_3=2;
V_4=30; V_L=-60; V_Ca=120; V_K=-84;
I=0;
c((1/cap)\emph{(I - g_Ca}0.5 * (1 + tanh((V - V_1) / V_2))\emph{(V-V_Ca)- g_K}W*(V-V_K)-g_L*(V-V_L)),
phi*(0.5 * (1 + tanh((V - V_3) / V_4)) - W)*cosh((V - V_3) / (2 * V_4)))
}
}
\details{
#'Modified from the original script pplane.r written by Daniel Kaplan,
Dept. of Mathematics, Macalester College, kaplan@macalester.edu

Modifications by S. Ellner for use in connection with the textbook
Dynamic Models in Biology by S.P. Ellner and J. Guckenheimer,
Princeton University Press (2006)

Further Modifications by PF. Bloniasz for use in connection with Math 1808
Biomathematics at Bowdoin College. The code has been condensed to only have
phase planes and nullclines. This script is being used under the presumption of
Fair Use copyright law for educational purposes. Questions should be directed to
patrick.bloniasz@gmail.com
}
